Moved Footer into components/layout
